* EOMutableKnownKeyDictionary.h/m ([-initWithObjects:forKeys:count:]): 	* EOFault.h/m ([-retainCount]): 	* EOKeyGlobalID.h/m ([-hash]): 	* EOMultiReaderLock.m ([-tryLockForReading]): 	* EOCheapArray.h/m ([-initWithObjects:count:],[initWithCapacity:]): 	* EOGlobalID.h/m ([-hash]): Use NS(U)Integer instead of int.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gdl2/trunk@30160 72102866-910b-0410-8b05-ffd578937521

+++ EOControl/EOCheapArray.h
@@ -45,7 +45,7 @@
 
 - (id)initWithArray: (NSArray *)array;
 - (id)initWithObjects: (id *)objects
-		count: (unsigned int)count;
+		count: (NSUInteger)count;
 - (void)release;
 - (unsigned int)retainCount;
 - (id)retain;
@@ -66,9 +66,9 @@
   id _immutableCopy;
 }
 
-- (id)initWithCapacity: (unsigned int)capacity;
+- (id)initWithCapacity: (NSUInteger)capacity;
 - (id)initWithObjects: (id *)objects
-		count: (unsigned int)count;
+		count: (NSUInteger)count;
 - (id)initWithArray: (NSArray *)array;
 - (void)dealloc;
 - (NSArray *)shallowCopy;
